chore: use parallel route for documents instead react async component wrapped in suspense
