+++ app/(header-default)/documents/[slug]/@actes/common.tsx
@@ -1,0 +1,8 @@
+import { EAdministration } from '#models/administrations/EAdministration';
+
+export const sectionInfo = {
+  title: 'Actes et statuts',
+  id: 'actes',
+  isProtected: true,
+  sources: [EAdministration.INPI],
+};
