+++ app/(header-default)/documents/[slug]/@carteProfessionnelleTP/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loader() {
+  return null;
+}
