+++ components/espace-agent-components/agent-wall/document.tsx
@@ -8,7 +8,6 @@ const AgentWallDocuments: React.FC<
     title: string;
     id: string;
     uniteLegale: IUniteLegale;
-    condition: boolean;
   }>
 > = ({ uniteLegale, id, title }) => (
   <AgentWall
